requirements.txt
# AmpyFin Val Model - requirements
# Runtime
python-dotenv>=1.0.1
requests>=2.32.3
curl-cffi>=0.6.0
yfinance
pandas>=2.2.2
numpy>=2.0.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
tenacity>=8.5.0

# Yahoo Finance session management (caching + rate limiting)
requests-cache>=1.2.1
requests-ratelimiter>=0.7.0
pyrate-limiter>=2.10.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
requests-mock>=1.12.1

# Lint/Format/Type (optional but recommended)
black>=24.8.0
ruff>=0.6.1
isort>=5.13.2
mypy>=1.11.2

# GUI (optional): install only if GUI_MODE=True
# PyQt5>=5.15.11
